Refactor: Split helpers and mappings into focused modules
